Only allow admins to add / remove users from an Organization

+++ src/routes/admin.php
@@ -9,37 +9,19 @@ use App\Http\Controllers\Admin\AdminOrganizationController;
 Route::group(['as' => 'admin.', 'prefix' => 'admin', 'middleware' => ['auth', 'verified', 'role:Admin,Organization Manager']], function () {
     Route::get('dashboard', function () {
         return Inertia::render('admin/Dashboard');
-    })
-    ->name('dashboard');
-
-    //Opportunity resource routes
-    // Route::resource('opportunities', AdminOpportunityController::class)
-    //     ->middleware(['auth','verified']);//->except(['index', 'show']);
+    })->name('dashboard');
 
     Route::group(['as' => 'opportunities.', 'prefix' => 'opportunities'], function () {
-         Route::get('', [AdminOpportunityController::class, 'index'])
-            ->name('index');
-        Route::get('create', [AdminOpportunityController::class, 'showCreate'])
-            ->name('create');
-        Route::post('', [AdminOpportunityController::class, 'store'])
-            ->name('store');
-        Route::get('{opportunity}', [AdminOpportunityController::class, 'show'])
-            ->name('show');
-        Route::post('{opportunity}/organizations/{organization}', [AdminOpportunityController::class, 'attachOrganization'])
-            ->name('organizations.attach');
-        Route::delete('{opportunity}/organizations/{organization}', [AdminOpportunityController::class, 'detachOrganization'])
-            ->name('organizations.detach');
-        Route::post('{opportunity}/tags', [AdminOpportunityController::class, 'addTag'])
-            ->name('tags.add');
-        Route::delete('{opportunity}/tags', [AdminOpportunityController::class, 'removeTag'])
-            ->name('tags.remove');
-        // Route::get('{opportunity}/edit', [AdminOpportunityController::class, 'showEdit'])
-        //     ->middleware(['auth','verified'])
-        //     ->name('showEdit');
-        Route::patch('{opportunity}', [AdminOpportunityController::class, 'update'])
-            ->name('update');
-        Route::delete('{opportunity}', [AdminOpportunityController::class, 'destroy'])
-            ->name('destroy');
+        Route::get('', [AdminOpportunityController::class, 'index'])->name('index');
+        Route::get('create', [AdminOpportunityController::class, 'showCreate'])->name('create');
+        Route::post('', [AdminOpportunityController::class, 'store']) ->name('store');
+        Route::get('{opportunity}', [AdminOpportunityController::class, 'show']) ->name('show');
+        Route::post('{opportunity}/organizations/{organization}', [AdminOpportunityController::class, 'attachOrganization'])->name('organizations.attach');
+        Route::delete('{opportunity}/organizations/{organization}', [AdminOpportunityController::class, 'detachOrganization'])->name('organizations.detach');
+        Route::post('{opportunity}/tags', [AdminOpportunityController::class, 'addTag'])->name('tags.add');
+        Route::delete('{opportunity}/tags', [AdminOpportunityController::class, 'removeTag'])->name('tags.remove');
+        Route::patch('{opportunity}', [AdminOpportunityController::class, 'update'])->name('update');
+        Route::delete('{opportunity}', [AdminOpportunityController::class, 'destroy'])->name('destroy');
     });
 
     Route::group(['as' => 'users.', 'prefix' => 'users', 'middleware' => ['role:Admin']], function () {
@@ -60,7 +42,7 @@ Route::group(['as' => 'admin.', 'prefix' => 'admin', 'middleware' => ['auth', 'v
         Route::get('{organization}', [AdminOrganizationController::class, 'show'])->name('show');
         Route::patch('{organization}', [AdminOrganizationController::class, 'update'])->name('update');
         Route::delete('{organization}', [AdminOrganizationController::class, 'destroy'])->name('destroy');
-        Route::post('{organization}/users/{user}', [AdminOrganizationController::class, 'attachUser'])->name('users.attach');
-        Route::delete('{organization}/users/{user}', [AdminOrganizationController::class, 'detachUser'])->name('users.detach');
+        Route::post('{organization}/users/{user}', [AdminOrganizationController::class, 'attachUser'])->name('users.attach')->middleware('role:Admin');
+        Route::delete('{organization}/users/{user}', [AdminOrganizationController::class, 'detachUser'])->name('users.detach')->middleware('role:Admin');
     });
 });
